This program manages a textfile database of songs.The program was designed with two classes, SongDatabase is the main class and Song is a class that represents an instance of a song.  The program keeps the songs in an ArrayList Each song is a csv String containing songName, songID, songDescription, songArtist, songAlbum, songPrice.
The data is read in the main class and placed in the ArrayList.  An Observable List of song names is created from the ArrayList to be used for the selections in the combobox.  Other methods in the program add songs, delete songs, edit songs and write the songs to a file after changes are made and on exit.

The GUI components are in the Stage methods.  All the layout labels, text field and combobox directives are contained there.