Create README.md
This program was designed with two classes, SongDatabase is the main class and contains the code to manipulate the database of song data and Song is a class that represents an instance of a song.  The program keeps the songs in a type Song ArrayList.   loadDBFromFile() reads the text file "database.txt" and creates the ArrayList.  If the text file cannot be found, the user is given the option to create an empty file and songs can be added from the GUI. The ArrayList is called as an argument by the LoadComboBox method which creates and ObservableList and is used to load the combobox.  When the combobox  called cbNames is clicked, the clickComboBox() method calls methods to load the song attributes into the GUI textboxes.  setButtonsAtStart() sets the Add, Delete, Cancel, Edit, Exit buttons correctly to enabled and disabled.  findSongFromCB() takes the song name from the combobox and is used to find the matching song in the database.  (What happens if the database contains more than one song with a name?  Need to test this.).   clickAddButton(), clickEditBtn() and clickDeleteBtn() reset the enable and disable of buttons to fit the action choice and also enable and disable or enable editing of the textboxes.  clickAcceptBtn() calls methods to implement the add, delete and edit functions and knows which to implement based on boolean values passed from add, delete and edit button methods.  validateAddedSong() checks that all the textbox fields are populated for the add and edit methods.  If all textboxes are not null, the method returns the newSong.  checkForDuplicates() takes the newSong and checks the database for a duplicate song which is defined as a song with the same name, artist and album as the newSong.   writeDBToFile uses PrintStream to write the ArrayList to the textfile.  clickCancleBtn() clears any changes made in the text boxes and returns the button statuses to the starting setting.  clickExitBtn() writes the ArrayList to a textfile and closes the GUI window.  setTheStage() contains the parameters to place the components of  the GUI in the pane.  This program uses an AnchorPane which allows each row of the pane to be individually composed so that the columnar components do not need to be aligned like in a GridPane.  The start() method is called by main() and calls the methods to load the database and create the combobox.  start() also contains the event handlers for the buttons and combobox.  This program uses a lambda expression for each button to create an anonymous class and implement a functional interface that contains one abstract method to handle the button click event.